test: continue splitting scanner tests by behavior
