+++ tests/test_device_scanner.py
@@ -1,7 +1,7 @@
 """Test device scanner for ThesslaGreen Modbus integration."""
 
 import logging
-from unittest.mock import AsyncMock, MagicMock, call, patch
+from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 from custom_components.thessla_green_modbus.const import (
@@ -32,498 +32,6 @@ HOLDING_REGISTERS = {r.name: r.address for r in get_registers_by_function(3)}
 INPUT_REGISTERS = {r.name: r.address for r in get_registers_by_function(4)}
 
 pytestmark = pytest.mark.asyncio
-
-
-async def test_read_holding_skips_after_failure():
-    """Holding registers are cached after a failed read."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=2)
-    mock_client = AsyncMock()
-
-    # Initial failing scan
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-            AsyncMock(side_effect=ModbusIOException("boom")),
-        ) as call_mock1,
-        patch("asyncio.sleep", AsyncMock()),
-    ):
-        result = await scanner._read_holding(mock_client, 168, 1)
-        assert result is None
-        assert call_mock1.await_count == scanner.retry
-
-    # Subsequent call should be skipped
-    with patch(
-        "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-        AsyncMock(),
-    ) as call_mock2:
-        result = await scanner._read_holding(mock_client, 168, 1)
-        assert result is None
-        call_mock2.assert_not_called()
-
-    assert 168 in scanner._failed_holding
-
-
-async def test_read_holding_skips_cached_failed_range_for_multi_register_read():
-    """Cached failed holding registers skip overlapping multi-register requests."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=2)
-    mock_client = AsyncMock()
-    scanner._failed_holding.update({170, 171, 172})
-
-    with patch(
-        "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-        AsyncMock(),
-    ) as call_mock:
-        result = await scanner._read_holding(mock_client, 170, 3)
-
-    assert result is None
-    call_mock.assert_not_called()
-    assert scanner.failed_addresses["modbus_exceptions"]["holding_registers"].issuperset(
-        {170, 171, 172}
-    )
-
-
-async def test_read_holding_exception_response(caplog):
-    """Exception responses should include the exception code in logs."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10)
-    mock_client = AsyncMock()
-
-    error_response = MagicMock()
-    error_response.isError.return_value = True
-    error_response.exception_code = 6
-
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-            AsyncMock(return_value=error_response),
-        ) as call_mock,
-        patch("asyncio.sleep", AsyncMock()),
-        caplog.at_level(logging.DEBUG),
-    ):
-        result = await scanner._read_holding(mock_client, 1, 1)
-
-    assert result is None
-    assert call_mock.await_count == scanner.retry
-    assert f"Exception code {error_response.exception_code}" in caplog.text
-
-
-async def test_read_input_exception_response_mentions_input_registers(caplog):
-    """Input exception responses should log the proper register type."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10)
-    mock_client = AsyncMock()
-
-    error_response = MagicMock()
-    error_response.isError.return_value = True
-    error_response.exception_code = 6
-
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-            AsyncMock(return_value=error_response),
-        ) as call_mock,
-        patch("asyncio.sleep", AsyncMock()),
-        caplog.at_level(logging.DEBUG),
-    ):
-        result = await scanner._read_input(mock_client, 1, 1)
-
-    assert result is None
-    assert call_mock.await_count == 1
-    assert "while reading input registers 1-1" in caplog.text
-
-
-async def test_read_holding_timeout_logging(caplog):
-    """Timeout errors should log a warning and a final error."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=2)
-    mock_client = AsyncMock()
-
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-            AsyncMock(side_effect=TimeoutError()),
-        ),
-        patch("asyncio.sleep", AsyncMock()),
-        caplog.at_level(logging.WARNING),
-    ):
-        result = await scanner._read_holding(mock_client, 1, 1)
-
-    assert result is None
-    warnings = [r.message for r in caplog.records if r.levelno == logging.WARNING]
-    assert any("Timeout reading holding 1" in msg for msg in warnings)
-    errors = [r.message for r in caplog.records if r.levelno == logging.ERROR]
-    assert any("Failed to read holding registers 1-1" in msg for msg in errors)
-
-
-async def test_read_holding_cancelled_request_logs_without_traceback(caplog):
-    """Cancelled pymodbus requests should not emit debug traceback spam."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=2)
-    mock_client = AsyncMock()
-    cancelled_exc = ModbusIOException("Request cancelled outside pymodbus.")
-
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-            AsyncMock(side_effect=cancelled_exc),
-        ),
-        patch("asyncio.sleep", AsyncMock()),
-        caplog.at_level(logging.DEBUG),
-    ):
-        result = await scanner._read_holding(mock_client, 4387, 1)
-
-    assert result is None
-    cancelled_records = [
-        record
-        for record in caplog.records
-        if "Cancelled reading holding registers 4387-4387" in record.getMessage()
-    ]
-    assert cancelled_records
-    assert all(record.exc_info is None for record in cancelled_records)
-
-
-@pytest.mark.parametrize(
-    "method, address",
-    [("_read_input", 1), ("_read_holding", 1)],
-)
-async def test_read_backoff_delay(method, address):
-    """Ensure exponential backoff delays between retries."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=3, backoff=0.1)
-    mock_client = AsyncMock()
-    if method == "_read_input":
-        mock_client.read_input_registers = AsyncMock(side_effect=ModbusIOException("boom"))
-        mock_client.read_holding_registers = AsyncMock(side_effect=ModbusIOException("boom"))
-    else:
-        mock_client.read_holding_registers = AsyncMock(side_effect=ModbusIOException("boom"))
-
-    sleep_mock = AsyncMock()
-    with patch(
-        "custom_components.thessla_green_modbus.modbus_helpers.asyncio.sleep",
-        sleep_mock,
-    ):
-        result = await getattr(scanner, method)(mock_client, address, 1)
-        assert result is None
-
-    assert [call.args[0] for call in sleep_mock.await_args_list] == [0.1, 0.2]
-
-
-@pytest.mark.parametrize(
-    "method, address",
-    [("_read_input", 1), ("_read_holding", 1)],
-)
-async def test_read_default_delay(method, address):
-    """Use default delay when backoff is not specified."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=3)
-    mock_client = AsyncMock()
-    if method == "_read_input":
-        mock_client.read_input_registers = AsyncMock(side_effect=ModbusIOException("boom"))
-        mock_client.read_holding_registers = AsyncMock(side_effect=ModbusIOException("boom"))
-    else:
-        mock_client.read_holding_registers = AsyncMock(side_effect=ModbusIOException("boom"))
-
-    sleep_mock = AsyncMock()
-    with patch(
-        "custom_components.thessla_green_modbus.modbus_helpers.asyncio.sleep",
-        sleep_mock,
-    ):
-        result = await getattr(scanner, method)(mock_client, address, 1)
-        assert result is None
-
-    non_zero = [c for c in sleep_mock.await_args_list if c.args[0] != 0]
-    assert non_zero == []
-
-
-@pytest.mark.parametrize(
-    "func, address",
-    [
-        (ThesslaGreenDeviceScanner._read_coil, 0),
-        (ThesslaGreenDeviceScanner._read_discrete, 0),
-    ],
-)
-async def test_read_binary_backoff_delay(func, address):
-    """Coil and discrete reads should respect configured backoff."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=3, backoff=0.1)
-    mock_client = AsyncMock()
-    if func is ThesslaGreenDeviceScanner._read_coil:
-        mock_client.read_coils = AsyncMock(side_effect=ModbusIOException("boom"))
-    else:
-        mock_client.read_discrete_inputs = AsyncMock(side_effect=ModbusIOException("boom"))
-
-    sleep_mock = AsyncMock()
-    with patch(
-        "custom_components.thessla_green_modbus.modbus_helpers.asyncio.sleep",
-        sleep_mock,
-    ):
-        result = await func(scanner, mock_client, address, 1)
-        assert result is None
-
-    assert [call.args[0] for call in sleep_mock.await_args_list] == [0.1, 0.2]
-
-
-@pytest.mark.parametrize(
-    "func, address",
-    [
-        (ThesslaGreenDeviceScanner._read_coil, 0),
-        (ThesslaGreenDeviceScanner._read_discrete, 0),
-    ],
-)
-async def test_read_binary_default_delay(func, address):
-    """Default backoff of zero should not delay retries."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=3)
-    mock_client = AsyncMock()
-    if func is ThesslaGreenDeviceScanner._read_coil:
-        mock_client.read_coils = AsyncMock(side_effect=ModbusIOException("boom"))
-    else:
-        mock_client.read_discrete_inputs = AsyncMock(side_effect=ModbusIOException("boom"))
-
-    sleep_mock = AsyncMock()
-    with patch(
-        "custom_components.thessla_green_modbus.modbus_helpers.asyncio.sleep",
-        sleep_mock,
-    ):
-        result = await func(scanner, mock_client, address, 1)
-        assert result is None
-
-    non_zero = [c for c in sleep_mock.await_args_list if c.args[0] != 0]
-    assert non_zero == []
-
-
-async def test_read_input_logs_warning_on_failure(caplog):
-    """Warn when input registers cannot be read after retries."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=2)
-    mock_client = AsyncMock()
-
-    caplog.set_level(logging.WARNING)
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-            AsyncMock(side_effect=ModbusIOException("boom")),
-        ) as call_mock,
-        patch("asyncio.sleep", AsyncMock()),
-    ):
-        result = await scanner._read_input(mock_client, 1, 1)
-        assert result is None
-        assert call_mock.await_count == scanner.retry
-
-    assert "Device does not expose register 1" in caplog.text
-
-
-async def test_read_input_skips_cached_failures():
-    """Input registers are cached after repeated failures."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=2)
-    mock_client = AsyncMock()
-
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-            AsyncMock(side_effect=ModbusIOException("boom")),
-        ) as call_mock,
-        patch("asyncio.sleep", AsyncMock()),
-    ):
-        result = await scanner._read_input(mock_client, 1, 1)
-        assert result is None
-        assert call_mock.await_count == scanner.retry
-
-    assert {1} <= scanner._failed_input
-
-    # Subsequent call should be skipped
-    with patch(
-        "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-        AsyncMock(),
-    ) as call_mock2:
-        result = await scanner._read_input(mock_client, 1, 1)
-        assert result is None
-        call_mock2.assert_not_called()
-
-
-async def test_read_input_skips_range_on_exception_response(caplog):
-    """Block failures with exception codes skip entire register range."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=2)
-    mock_client = AsyncMock()
-
-    error_response = MagicMock()
-    error_response.isError.return_value = True
-    error_response.exception_code = 2
-
-    caplog.set_level(logging.WARNING)
-    with patch(
-        "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-        AsyncMock(return_value=error_response),
-    ) as call_mock:
-        result = await scanner._read_input(mock_client, 256, 3)
-
-    scanner._log_skipped_ranges()
-
-    assert result is None
-    assert call_mock.await_count == 1
-    assert set(range(256, 259)) <= scanner._failed_input
-    assert "Skipping unsupported input registers 256-258 (exception code 2)" in caplog.text
-
-    # Further reads within the range should be skipped without new calls
-    with patch(
-        "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-        AsyncMock(),
-    ) as call_mock2:
-        result = await scanner._read_input(mock_client, 257, 1)
-        assert result is None
-        call_mock2.assert_not_called()
-
-
-async def test_read_holding_skips_range_on_exception_response(caplog):
-    """Block failures with exception codes skip entire holding register range."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=2)
-    mock_client = AsyncMock()
-
-    error_response = MagicMock()
-    error_response.isError.return_value = True
-    error_response.exception_code = 2
-
-    caplog.set_level(logging.WARNING)
-    with patch(
-        "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-        AsyncMock(return_value=error_response),
-    ) as call_mock:
-        result = await scanner._read_holding(mock_client, 512, 2)
-
-    scanner._log_skipped_ranges()
-
-    assert result is None
-    assert call_mock.await_count == 1
-    assert (512, 513) in scanner._unsupported_holding_ranges
-    assert "Skipping unsupported holding registers 512-513 (exception code 2)" in caplog.text
-
-    # Further reads within the range should be skipped without new calls
-    with patch(
-        "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-        AsyncMock(),
-    ) as call_mock2:
-        result = await scanner._read_holding(mock_client, 513, 1)
-        assert result is None
-        call_mock2.assert_not_called()
-
-
-async def test_read_holding_skip_cache_reads_unsupported_range():
-    """Single reads are attempted even if range was marked unsupported."""
-    scanner = await ThesslaGreenDeviceScanner.create("host", 502, 10)
-    mock_client = AsyncMock()
-    scanner._unsupported_holding_ranges[(512, 513)] = 1
-    scanner._failed_holding.update({512, 513})
-
-    response = MagicMock()
-    response.isError.return_value = False
-    response.registers = [123]
-
-    with patch(
-        "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-        AsyncMock(return_value=response),
-    ) as call_mock:
-        result = await scanner._read_holding(mock_client, 512, 1, skip_cache=True)
-
-    assert result == [123]
-    call_mock.assert_awaited_once()
-
-
-async def test_single_read_clears_failed_holding_cache():
-    """Successful single reads remove addresses from cached failure ranges."""
-    scanner = await ThesslaGreenDeviceScanner.create("host", 502, 10)
-    client = AsyncMock()
-    scanner._mark_holding_unsupported(768, 770, 1)
-    scanner._failed_holding.update({768, 769, 770})
-
-    response = MagicMock()
-    response.isError.return_value = False
-    response.registers = [55]
-
-    with patch(
-        "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-        AsyncMock(return_value=response),
-    ) as call_mock:
-        result = await scanner._read_holding(client, 769, 1, skip_cache=True)
-
-    assert result == [55]
-    call_mock.assert_awaited_once()
-    assert 769 not in scanner._failed_holding
-    assert (768, 768) in scanner._unsupported_holding_ranges
-    assert (770, 770) in scanner._unsupported_holding_ranges
-    assert all(not (start <= 769 <= end) for start, end in scanner._unsupported_holding_ranges)
-
-
-async def test_log_skipped_ranges_no_overlap(caplog):
-    """Logged skipped ranges should not contain overlaps."""
-    scanner = await ThesslaGreenDeviceScanner.create("host", 502, 10)
-    scanner._mark_holding_unsupported(1024, 1028, 1)
-    scanner._mark_holding_unsupported(1026, 1030, 2)
-
-    with caplog.at_level(logging.WARNING):
-        scanner._log_skipped_ranges()
-
-    assert "1024-1025 (exception code 1)" in caplog.text
-    assert "1026-1030 (exception code 2)" in caplog.text
-    assert "1024-1028" not in caplog.text
-
-
-async def test_read_input_logs_once_per_skipped_range(caplog):
-    """Only one log message is emitted per skipped register range."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.3.17", 8899, 10, retry=2)
-    mock_client = AsyncMock()
-    scanner._failed_input.update({1, 2, 3})
-
-    caplog.set_level(logging.DEBUG)
-    for addr in range(1, 4):
-        result = await scanner._read_input(mock_client, addr, 1)
-        assert result is None
-
-    messages = [
-        record.message
-        for record in caplog.records
-        if "Skipping cached failed input registers" in record.message
-    ]
-    assert messages == ["Skipping cached failed input registers 1-3"]
-
-
-async def test_read_holding_exponential_backoff(caplog):
-    """Ensure exponential backoff and error reporting when device fails to respond."""
-    scanner = await ThesslaGreenDeviceScanner.create("host", 502, 10, retry=3, backoff=0.5)
-    client = AsyncMock()
-
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-            AsyncMock(side_effect=ModbusException("boom")),
-        ),
-        patch(
-            "custom_components.thessla_green_modbus.modbus_helpers.asyncio.sleep",
-            AsyncMock(),
-        ) as sleep_mock,
-        caplog.at_level(logging.WARNING),
-    ):
-        result = await scanner._read_holding(client, 1, 1)
-
-    assert result is None
-    assert sleep_mock.await_args_list == [call(0.5), call(1.0)]
-    assert any(
-        "Failed to read holding registers 1-1" in record.message for record in caplog.records
-    )
-
-
-async def test_read_holding_returns_none_on_modbus_error():
-    """A Modbus error response should return None without raising."""
-    scanner = await ThesslaGreenDeviceScanner.create("host", 502, 10)
-    client = AsyncMock()
-    response = MagicMock()
-    response.isError.return_value = True
-    call_modbus = AsyncMock(return_value=response)
-
-    with (
-        patch(
-            "custom_components.thessla_green_modbus.scanner.io_core._call_modbus",
-            call_modbus,
-        ),
-        patch(
-            "custom_components.thessla_green_modbus.modbus_helpers.asyncio.sleep",
-            AsyncMock(),
-        ),
-    ):
-        result = await scanner._read_holding(client, 1, 1)
-
-    assert result is None
-    assert call_modbus.await_count == scanner.retry
 
 
 async def test_scan_device_success_dynamic():
